+++ example_web/src/lib.rs
@@ -13,8 +13,6 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 pub fn start(canvas_id: &str) -> Result<(), wasm_bindgen::JsValue> {
     let app = example_app::ExampleApp::default();
-    let backend = egui_web::WebBackend::new(canvas_id)?;
-    let runner = egui_web::AppRunner::new(backend, Box::new(app))?;
-    egui_web::start(runner)?;
+    egui_web::start(canvas_id, Box::new(app))?;
     Ok(())
 }